Anim fixes
