+++ Assets/Animation/Guard/GuardAnim.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7277527ad9ebd2441865d2893880afbb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
