+++ Assets/Animation/Guard/GuardIdle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 318454fc713cff6408ce0ace0aca1285
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
