+++ Assets/Animation/Guard/GuardWalk.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 02d2fbd1c1fdfce46ac442c20da16863
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
