+++ Assets/Animation/Guard/Take 001.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: befcb5aae88426f49a51ccc3e23fb5ad
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
