+++ Assets/Scripts/AnimController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 906eda9f94020184698629802639f06a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
