+++ Assets/Scripts/AnimController/AnimController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 821fc798e89f8cb4cb0dd8044dae1764
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
