+++ Assets/Scripts/AnimController/GuardAnimController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4f1f9756a22602e4992cd0f4fe9bb6ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
